+++ app/Core/Httpr.ts
@@ -1,5 +1,5 @@
 import * as _ from 'lodash';
-import {Map} from '../Type/Map';
+import {StringMap} from '../Type/StringMap';
 import {HttpMethod} from '../Type/HttpMethod';
 import {PlainObject} from '../Type/PlainObject';
 import {HttpMethods} from '../Enum/HttpMethods';
@@ -52,89 +52,104 @@ export class Httpr {
   /**
    * Perform a GET request.
    *
+   * @see {@link Httpr.request}
+   *
    * @param {string} url - url of resource to request.
    * @param {PlainObject} params - hash of additional query parameters.
-   * @param {Map<string>} headers - hash of request headers to set.
+   * @param {StringMap} headers - hash of request headers to set.
    * @returns {Promise<HttpResponse>} promise handler.
    */
-  public get(url: string, params?: PlainObject, headers?: Map<string>): Promise<HttpResponse> {
+  public get(url: string, params?: PlainObject, headers?: StringMap): Promise<HttpResponse> {
     return this.request(HttpMethods.GET, url, params, headers);
   }
 
   /**
    * Perform a POST request.
    *
+   * @see {@link Httpr.request}
+   *
    * @param {string} url - url of resource to request.
    * @param {PlainObject} params - hash of additional query parameters.
    * @param {*} body - request body data.
-   * @param {Map<string>} headers - hash of request headers to set.
+   * @param {StringMap} headers - hash of request headers to set.
    * @returns {Promise<HttpResponse>} promise handler.
    */
-  public post(url: string, params?: PlainObject, body?: any, headers?: Map<string>): Promise<HttpResponse> {
+  public post(url: string, params?: PlainObject, body?: any, headers?: StringMap): Promise<HttpResponse> {
     return this.request(HttpMethods.POST, url, params, headers, body);
   }
 
   /**
    * Perform a PUT request.
    *
+   * @see {@link Httpr.request}
+   *
    * @param {string} url - url of resource to request.
    * @param {PlainObject} params - hash of additional query parameters.
    * @param {*} body - request body data.
-   * @param {Map<string>} headers - hash of request headers to set.
+   * @param {StringMap} headers - hash of request headers to set.
    * @returns {Promise<HttpResponse>} promise handler.
    */
-  public put(url: string, params?: PlainObject, body?: any, headers?: Map<string>): Promise<HttpResponse> {
+  public put(url: string, params?: PlainObject, body?: any, headers?: StringMap): Promise<HttpResponse> {
     return this.request(HttpMethods.PUT, url, params, headers, body);
   }
 
   /**
    * Perform a DELETE request.
    *
+   * @see {@link Httpr.request}
+   *
    * @param {string} url - url of resource to request.
    * @param {PlainObject} params - hash of additional query parameters.
-   * @param {Map<string>} headers - hash of request headers to set.
+   * @param {StringMap} headers - hash of request headers to set.
    * @returns {Promise<HttpResponse>} promise handler.
    */
-  public del(url: string, params?: PlainObject, headers?: Map<string>): Promise<HttpResponse> {
+  public del(url: string, params?: PlainObject, headers?: StringMap): Promise<HttpResponse> {
     return this.request(HttpMethods.DELETE, url, params, headers);
   }
 
   /**
    * Perform an OPTIONS request.
    *
+   * @see {@link Httpr.request}
+   *
    * @param {string} url - url of resource to request.
    * @param {PlainObject} params - hash of additional query parameters.
-   * @param {Map<string>} headers - hash of request headers to set.
+   * @param {StringMap} headers - hash of request headers to set.
    * @returns {Promise<HttpResponse>} promise handler.
    */
-  public options(url: string, params?: PlainObject, headers?: Map<string>): Promise<HttpResponse> {
+  public options(url: string, params?: PlainObject, headers?: StringMap): Promise<HttpResponse> {
     return this.request(HttpMethods.OPTIONS, url, params, headers);
   }
 
   /**
    * Perform a PATCH request.
    *
+   * @see {@link Httpr.request}
+   *
    * @param {string} url - url of resource to request.
    * @param {PlainObject} params - hash of additional query parameters.
-   * @param {Map<string>} headers - hash of request headers to set.
-   * @returns {Promise<HttpResponse>} promise handler.
+   * @param {StringMap} headers - hash of request headers to set.
+   * @returns {Promise<HttpResponse>} Promise handler.
    */
-  public patch(url: string, params?: PlainObject, headers?: Map<string>): Promise<HttpResponse> {
+  public patch(url: string, params?: PlainObject, headers?: StringMap): Promise<HttpResponse> {
     return this.request(HttpMethods.PATCH, url, params, headers);
   }
 
   /**
    * Perform an HTTP Request.
    *
+   * @see {@link HttprStatic}
+   * @see {@link HttprProvider}
+   *
    * @param {HttpMethod} method - http method to use.
    * @param {string} url - url of resource to request.
    * @param {PlainObject} params - hash of additional query parameters.
-   * @param {Map<string>} headers - hash of request headers to set.
+   * @param {StringMap} headers - hash of request headers to set.
    * @param {*} body - request body data.
    * @returns {Promise<HttpResponse>} promise handler.
    */
   public request(method: HttpMethod, url: string, params?: PlainObject,
-                 headers?: Map<string>, body?: any): Promise<HttpResponse> {
+                 headers?: StringMap, body?: any): Promise<HttpResponse> {
     return this.config.provider
     .request(HttprStatic.build(this, method, url, params, headers, body))
     .then((data) => HttprStatic.onSuccess(this, data))