+++ app/Enum/HttpHeaders.ts
@@ -1,5 +1,9 @@
 /**
  * Http Header constants.
+ *
+ * Bundled headers:
+ * - accept
+ * - content-type
  */
 export class HttpHeaders {
   public static ACCEPT = 'accept';