+++ app/Enum/MediaTypes.ts
@@ -1,5 +1,9 @@
 /**
  * Media Type constants.
+ *
+ * Bundled Media Types:
+ * - application/json
+ * - application/x-www-form-urlencoded
  */
 export class MediaTypes {
   public static APPLICATION_JSON = 'application/json';