+++ app/Type/HttpRequestSettings.ts
@@ -1,4 +1,4 @@
-import {Map} from '../Type/Map';
+import {StringMap} from './StringMap';
 import {HttpMethod} from './HttpMethod';
 import {PlainObject} from './PlainObject';
 
@@ -24,7 +24,7 @@ export type HttpRequestSettings = {
   /**
    * Request headers.
    */
-  headers: Map<string>;
+  headers: StringMap;
 
   /**
    * Request body.