+++ app/Type/StringMap.ts
@@ -1,0 +1,6 @@
+/**
+ * String hash map.
+ */
+export type StringMap = {
+  [key: string]: string;
+}